elk/python_apps/server/requirements.txt
Flask==3.1.3
prometheus-flask-exporter==0.23.2
requests==2.32.5
opentelemetry-api==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-instrumentation-flask==0.61b0
opentelemetry-instrumentation-requests==0.61b0
opentelemetry-exporter-otlp==1.40.0
